Fix handling of javadoc and source jars to make them not a part of the default configuration.
Being part of the default configuration, these jars were being exported to dependency submodules. Spock seems to have issues compiling specs when a source jar is on the compile path. I've moved the javadoc/source jars into their own configuration which is not exported.

Unfortunately, the signing plugin can't deal with configurations other than the default. So I've forked that and pulled our fork in to the build with a hack to work with our configuration as well as the default.