GEB-153 - in ConfigurationLoader#getConf(null) is called, treat it like Configuration.getConf()
This means that getConf(null) will also look for config _classes_ as well. The test adapter integrations all use this variant to allow user's an easy way of using a custom config file per test.

This is getting a bit messy, but now is not the right time for a big change.