Fix a bug under Groovy 3 which caused PageInstanceNotInitializedException when container was used as a content element name.
When handling missing methods, Groovy 3 now attempts to call the "call" method on a property which has the same name as the missing method being called if there is one. As part of evaluating the content closure method calls are dispatched to the Page class (which is the owner of the closure) which has a methodMissing() implementation delegating to PageContentSupport. At the time of the method call dispatch from the content closure evaluation the page is not yet initialised hence the exception. The solution is to make sure that the container property in PageContentSupport does not handle missing methods and make Groovy aware that there is no such method as "call".

Fixes geb/issues#640.